set dir framework
